Create cache system
