+++ web/app.fluidity.money/package.json
@@ -70,6 +70,7 @@
     "socket.io": "^4.5.2",
     "socket.io-client": "^4.5.2",
     "subscriptions-transport-ws": "^0.9.19",
+    "swr": "^1.3.0",
     "tinycolor2": "^1.4.2",
     "toml": "^3.0.0",
     "web3": "^1.8.0",
